feat(ai): add WebSocket transport to openai-responses provider
Extract shared WebSocket utilities from openai-codex-responses into
openai-websocket.ts and add WebSocket transport support to the standard
openai-responses provider. Fix buildBaseOptions to forward the transport
option through the full pipeline.
